fix: :truck: Moved music/lib.ts to lib/music.ts
Issue: #131
