+++ src/lib/music.ts
